recomendation/preprocessing/requirements.txt
python-dotenv
requests
pycountry
tqdm